Don't delete phenotype if we have pending data
Issue: IBP-2349 / IBP-2409 / IBP-2427
Reviewer: None

+++ src/main/webapp/WEB-INF/static/js/trialmanager/subobservations/subObservationSet.js
@@ -402,8 +402,12 @@
 
 								if (cellData.observationId) {
 									if (!value && !$scope.isPendingView) {
-										return datasetService.deleteObservation(subObservationSet.id, rowData.observationUnitId,
-											cellData.observationId);
+										if (cellData.draftValue) {
+                                            value = null;
+										} else {
+											return datasetService.deleteObservation(subObservationSet.id, rowData.observationUnitId,
+												cellData.observationId);
+										}
 									}
 
 									return confirmOutOfBoundData(value, columnData).then(function (doContinue) {
